+++ src/components/layout/layout.tsx
@@ -1,69 +1,49 @@
 import { Outlet, useLocation} from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../const';
-import { getAuthorizationStatus } from '../../utils/authtorization-status';
+import { AppRoute} from '../const';
+// import { getAuthorizationStatus } from '../../utils/authtorization-status';
+import Header from '../header';
 
-const getLayoutState = (pathname: AppRoute) => {
-  let rootClassName = '';
-  let linkClassName = '';
-  let shouldRenderUser = true;
+// const getLayoutState = (pathname: AppRoute) => {
+//   let rootClassName = '';
+//   let linkClassName = '';
+//   let shouldRenderUser = true;
+//   let shouldRenderFooter = false;
 
-  if (pathname === AppRoute.Main) {
-    rootClassName = 'page--gray page--main';
-    linkClassName = 'header__logo-link--active';
-  } else if (pathname === AppRoute.Login) {
-    rootClassName = 'page--gray page--login';
-    shouldRenderUser = false;
-  }
+//   if (pathname === AppRoute.Main) {
+//     rootClassName = 'page--gray page--main';
+//     linkClassName = 'header__logo-link--active';
+//   } else if (pathname === AppRoute.Login) {
+//     rootClassName = 'page--gray page--login';
+//     shouldRenderUser = false;
+//   } else if (pathname === AppRoute.Favorites) {
+//     shouldRenderFooter = true;
+//   }
 
-  return {rootClassName, linkClassName, shouldRenderUser};
-};
+//   return {rootClassName, linkClassName, shouldRenderUser, shouldRenderFooter};
+// };
 
 export default function Layout() {
+  let shouldRenderFooter = false;
   const {pathname} = useLocation();
-  const {rootClassName, linkClassName, shouldRenderUser} = getLayoutState(pathname as AppRoute);
-  const authorizationStatus = getAuthorizationStatus();
+  // console.log(pathname);
+  // console.log(AppRoute.Favorites)
+  // const {rootClassName, linkClassName, shouldRenderUser, shouldRenderFooter} = getLayoutState(pathname as AppRoute);
+  // const authorizationStatus = getAuthorizationStatus();
+  if (pathname === AppRoute.Favorites) {
+    shouldRenderFooter = true;
+  }
 
   return (
-    <div className={`page${rootClassName}`}>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className={`header__logo-link${linkClassName}`}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </a>
-            </div>
-            {
-              shouldRenderUser ? (
-                <nav className="header__nav">
-                  <ul className="header__nav-list">
-                    <li className="header__nav-item user">
-                      <a className="header__nav-link header__nav-link--profile" href="#">
-                        <div className="header__avatar-wrapper user__avatar-wrapper">
-                        </div>
-                        {authorizationStatus === AuthorizationStatus.Authorization ? (
-                          <>
-                            <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                            <span className="header__favorite-count">3</span>
-                          </>) : <span className="header__login">Sign in</span>}
-                      </a>
-                    </li>
-
-                    {authorizationStatus === AuthorizationStatus.Authorization ? (
-                      <li className="header__nav-item">
-                        <a className="header__nav-link" href="#">
-                          <span className="header__signout">Sign out</span>
-                        </a>
-                      </li>
-                    ) : null}
-                  </ul>
-                </nav>
-              ) : null
-            }
-          </div>
-        </div>
-      </header>
+    <>
+      <Header pathname = {pathname} />
       <Outlet />
-    </div>
+      {shouldRenderFooter ? (
+        <footer className="footer container">
+          <a className="footer__logo-link" href="main.html">
+            <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
+          </a>
+        </footer>
+      ) : null}
+    </>
   );
 }
